+++ src/app.ts
@@ -31,3 +31,7 @@ app.register(fastifySwagger, optsSwagger);
 app.register(userRouter);
 app.register(boardRouter);
 app.register(taskRouter);
+
+// throw Error('Oops!');
+
+// Promise.reject(Error('Oops!'));
