+++ src/resources/boards/board.routes/board.schema.ts
@@ -1,4 +1,4 @@
-import boardHandlerRouters from './board.handler.routers';
+import boardHandlerRouters from './board.controller';
 
 const board = {
   type: 'object',
