+++ src/resources/tasks/task.routes/task.controller.ts
@@ -3,7 +3,7 @@ import taskService from '../task.service';
 import statusCode from '../../../common/status.code';
 import Task from '../task.model';
 import { ITask } from '../../../common/type';
-import { logger, logCollect } from '../../../common/logger';
+import { logger, getLogObject } from '../../../logging/logger';
 import { NotFoundError } from '../../../errors/custom.errors';
 
 type FastifyRequestTask = FastifyRequest<{
@@ -24,7 +24,7 @@ type FastifyRequestTask = FastifyRequest<{
 const getTasksAllRouter = async (_: FastifyRequest, reply: FastifyReply) => {
   const tasks = await taskService.getTasksAllService();
   reply.code(statusCode.OK).send(tasks);
-  logger.info(logCollect(_, reply));
+  logger.info(getLogObject(_, reply));
 };
 
 /**
@@ -47,7 +47,7 @@ const getTaskIdRouter = async (
   if (await taskService.getTaskIdService(taskId)) {
     const task = await taskService.getTaskIdService(taskId);
     reply.code(statusCode.OK).send(task);
-    logger.info(logCollect(request, reply));
+    logger.info(getLogObject(request, reply));
   } else {
     throw new NotFoundError('Not found task');
   }
@@ -72,7 +72,7 @@ const addTaskRouter = async (
   const task: ITask = new Task(request.body);
   await taskService.addTaskService(task);
   reply.code(statusCode.CREATED).send(task);
-  logger.info(logCollect(request, reply));
+  logger.info(getLogObject(request, reply));
 };
 
 /**
@@ -94,7 +94,7 @@ const updateTaskRouter = async (
     const updTask: ITask = new Task(request.body, taskId);
     await taskService.updateTaskService(taskId, updTask);
     reply.code(statusCode.OK).send(updTask);
-    logger.info(logCollect(request, reply));
+    logger.info(getLogObject(request, reply));
   } else {
     throw new NotFoundError('Not found task');
   }
@@ -120,7 +120,7 @@ const deleteTaskRouter = async (
   if (await taskService.getTaskIdService(taskId)) {
     await taskService.deleteTaskService(taskId);
     reply.code(statusCode.NO_CONTENT);
-    logger.info(logCollect(request, reply));
+    logger.info(getLogObject(request, reply));
   } else {
     throw new NotFoundError('Not found task');
   }
