+++ src/resources/users/user.routes/user.router.ts
@@ -1,6 +1,6 @@
 import { FastifyPluginAsync, FastifyRequest, FastifyReply } from 'fastify';
 import userSchema from './user.schema';
-import { logger, logCollect } from '../../../common/logger';
+import { logger, getLogObject } from '../../../logging/logger';
 import statusCode from '../../../common/status.code';
 
 /**
@@ -11,12 +11,12 @@ import statusCode from '../../../common/status.code';
 const userRoutes: FastifyPluginAsync = async (app): Promise<void> => {
   app.get('/*', (_: FastifyRequest, reply: FastifyReply) => {
     reply.status(statusCode.NOT_FOUND).send('Not Found URL');
-    logger.warn(logCollect(_, reply));
+    logger.warn(getLogObject(_, reply));
   });
 
   app.get('/', async (_: FastifyRequest, reply: FastifyReply) => {
     reply.send({ Message: 'Service: is running!' });
-    logger.info(logCollect(_, reply));
+    logger.info(getLogObject(_, reply));
   });
 
   app.get('/users', userSchema.getUsersOpts);
